Install Shorebird CLI in your Bitrise Workflow
This step enables the usage of shorebird in your Bitrise workflows.
It not only installs the CLI but attempts to synchronize Flutter
versions to ensure compatibility.

If Flutter is not installed then Shorebird will use the latest
stable Flutter version. It is highly recommended that you still
install Flutter either through the Bitrise Flutter install step
or the FVM Step.

Shorebird will attempt to switch to the Flutter versions that your
project uses.